WaveManager refactor
