+++ Assets/Scenes/SampleScene.unity
@@ -12527,6 +12527,7 @@ GameObject:
   m_Component:
   - component: {fileID: 513389693}
   - component: {fileID: 513389694}
+  - component: {fileID: 513389695}
   m_Layer: 0
   m_Name: MainWaveManager
   m_TagString: Untagged
@@ -12564,8 +12565,24 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 8b7fabbde93a50c4f920b876d5c5f464, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  waveCombo: 0123
+--- !u!114 &513389695
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 513389692}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 06cfc9d57bddcf84c92916991b549746, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  waveManager: {fileID: 513389694}
   BuildingCount: 0
+  meleeEnemiesSpawnScaleFactor: 0
+  rangedEnemiesSpawnScaleFactor: 0
+  enemiesBetterTypeScaleFactor: 0
+  waveCombo: 
   spawners:
   - {fileID: 245971837}
   - {fileID: 1759539623}
