+++ Assets/Scripts/Menu/PauseMenu.cs
@@ -17,14 +17,14 @@ public class PauseMenu : MonoBehaviour
     public GameObject Dead;
     public GameObject SureGameover;
     public GameObject progressText;
-    private WaveManager WaveManager;
+    private WaveManagerSubscriber WaveManager;
 
 
     public GameObject GUI;
 
     void Awake()
     {
-        WaveManager = FindObjectOfType<WaveManager>();
+        WaveManager = FindObjectOfType<WaveManagerSubscriber>();
         pauseMenuUI.SetActive(false);
         GameIsPaused = false;
         GUI = GameObject.Find("GUI");
